fix faq title mb, bigger images in footer, footer content, add intro page, add homepage
