+++ locales/en.yml
@@ -34,22 +34,15 @@ not-found: Not found
 section02:
   title: Distribution
   desc: |
-   We have committed to a [b]fair[/b] and [b]easy to understand[/b] distribution method.
-   There will be only [b]two phases[/b] for mint: one for our 
-   [u]Whitelist holders[/u], given to those who have been integral in growing and contributing to our 
-   amazing project and a [u]public mint[/u]. The mint prices as of now are 
-   [b].08 ETH for public[/b] and for Whitelist there is a small discount of 
-   [b].07 ETH[/b]. Our [b]highly optimized contract[/b] created by our amazing developer should result in 
-   [b]low gas fees[/b] and a [b]seamless and effective minting process[/b].
-
-   Our mint will take place sometime in Q1 of 2024!
+   	We have committed to a fair and easy to understand distribution method.There will be only two phases for mint: one for our Whitelist holders, given to those who have been integral in growing and contributing to our amazing project and a public mint.
+  desc2: |
+    The mint prices as of now are .08 ETH for public and for Whitelist there is a discount of .07 ETH. Highly optimized contract created by our amazing developer should result in low gas fees and a seamless and effective minting process. Our mint will take place sometime in Q1 of 2024!
   button: Mint Now
 section03:
   title: Roadmap
   desc: |
     Keep an eye on the ⁠📢announcements channel for new and upcoming utilities!.
   button: Discord
-
 section04:
   wldate: TBA
   weprice: .07 ETH
@@ -58,7 +51,8 @@ section04:
 section05:
   title: Frequently Asked Questions
 footer:
-  title: HaHa Hyenas NFT Collection
+  title: HaHa Hyenas
+  title2: Our Links
   year: 2023
-  rights: All Rights Reserved
+  rights: All Rights Reserved
