+++ src/layouts/index.vue
@@ -1,0 +1,10 @@
+<!-- eslint-disable vue/multi-word-component-names -->
+<template>
+    <main text="center gray-700 dark:gray-200" px-8 > 
+      <RouterView />
+      <div mx-auto mt-5 text-center text-sm opacity-50>
+        [Index Layout]
+      </div> 
+    </main>
+  </template>
+
