+++ src/styles/button01.less
@@ -4,8 +4,7 @@
 @white-light-opacity: 0.9;
 
 .button-container {
-  position: relative;
-  display: inline-block;
+  position: relative; 
 
   .custom-btn {
     background: @background-color;
