+++ src/styles/style.less
@@ -56,12 +56,11 @@ li {
 }
 
 
-nav, #section01, #section03, #section04 {
+nav, #section01, #section03 {
     .hahaha-text{
         span {
             display: none;
         }
-
     }
 }
